unified workspace dependencies, added versions
