+++ crates/wick/wick-component-cli/Cargo.toml
@@ -15,13 +15,13 @@ grpc = ['wick-rpc', 'tonic/tls', 'wick-invocation-server', 'async-stream']
 mesh = ['wick-rpc', 'async-stream']
 
 [dependencies]
-logger = { path = "../logger" }
-wick-rpc = { path = "../wick-rpc", optional = true }
-wick-packet = { path = "../wick-packet" }
-wick-invocation-server = { path = "../wick-invocation-server", optional = true }
-cfg-if = "1.0"
+logger = { workspace = true }
+wick-rpc = { workspace = true, optional = true }
+wick-packet = { workspace = true }
+wick-invocation-server = { workspace = true, optional = true }
+cfg-if = { workspace = true }
 tonic = { workspace = true, optional = true }
-tonic-reflection = { version = "0.6", default-features = false, optional = true }
+tonic-reflection = { workspace = true, default-features = false, optional = true }
 tokio = { workspace = true, features = [
   "macros",
   "fs",
@@ -30,7 +30,7 @@ tokio = { workspace = true, features = [
   "rt-multi-thread",
 ] }
 tokio-stream = { workspace = true }
-async-stream = { version = "0.3.2", optional = true }
+async-stream = { workspace = true, optional = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 tracing = { workspace = true }
@@ -39,10 +39,10 @@ clap = { workspace = true, features = ["derive", "env"] }
 uuid = { workspace = true, features = ["v4"] }
 
 [dev-dependencies]
-logger = { path = "../logger" }
-test-logger = { path = "../../misc/test-logger" }
-test-native-collection = { path = "../../integration/test-native-collection" }
+logger = { workspace = true }
+test-logger = { workspace = true }
+test-native-collection = { workspace = true }
 once_cell = { workspace = true }
-reqwest = { version = "0.11", features = ["json"] }
+reqwest = { workspace = true, features = ["json"] }
 anyhow = { workspace = true }
 serde_json = { workspace = true }
