+++ crates/wick/wick-component-codegen/Cargo.toml
@@ -5,16 +5,15 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
-quote = "1.0.26"
 serde_yaml = { workspace = true }
-syn = "1.0.109"
-wick-config = { path = "../wick-config" }
-wick-interface-types = { path = "../wick-interface-types" }
-proc-macro2 = "1.0"
+wick-config = { workspace = true }
+wick-interface-types = { workspace = true }
 anyhow = { workspace = true }
-heck = "0.4"
+quote = { workspace = true }
+syn = { workspace = true }
+proc-macro2 = { workspace = true }
+heck = { workspace = true }
 
 [dev-dependencies]
-anyhow = { workspace = true }
-logger = { path = "../logger" }
-test-logger = { path = "../../misc/test-logger" }
+logger = { workspace = true }
+test-logger = { workspace = true }
