+++ crates/wick/wick-grpctar/Cargo.toml
@@ -10,17 +10,17 @@ publish = false
 [dependencies]
 thiserror = { workspace = true }
 tracing = { workspace = true }
-serde_json = "1"
+serde_json = { workspace = true }
 serde_yaml = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
-tar = "0.4.38"
-wick-interface-types = { path = "../wick-interface-types" }
-wick-wascap = { path = "../wick-wascap" }
+tar = { workspace = true }
+wick-interface-types = { workspace = true }
+wick-wascap = { workspace = true }
 
 [dev-dependencies]
 anyhow = { workspace = true }
-test-logger = { path = "../../misc/test-logger" }
-logger = { path = "../logger" }
+test-logger = { workspace = true }
+logger = { workspace = true }
 clap = { workspace = true, features = ["derive", "env"] }
 async-recursion = { workspace = true }
 env_logger = { workspace = true }
