+++ crates/wick/wick-interface-types/Cargo.toml
@@ -18,14 +18,14 @@ yaml = ["serde_yaml"]
 
 [dependencies]
 serde = { workspace = true, features = ["derive"] }
-serde_repr = "0.1"
-nom = { version = "7.1", optional = true }
+serde_repr = { workspace = true }
+nom = { workspace = true, optional = true }
 serde_yaml = { workspace = true, optional = true }
 
 [dev-dependencies]
 anyhow = { workspace = true }
-test-log = "0.2.8"
+test-log = { workspace = true }
 env_logger = { workspace = true }
-pretty_assertions = "1.2"
+pretty_assertions = { workspace = true }
 serde_json = { workspace = true }
 serde_yaml = { workspace = true }
