+++ crates/wick/wick-loader-utils/Cargo.toml
@@ -8,7 +8,7 @@ publish = false
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
-wick-oci-utils = { path = "../../wick/wick-oci-utils" }
+wick-oci-utils = { workspace = true }
 tracing = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["fs"] }
