+++ crates/wick/wick-wascap/Cargo.toml
@@ -8,15 +8,15 @@ publish = false
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
-wascap = "0.8"
 serde = { workspace = true, features = ["derive"] }
-wick-interface-types = { path = "../wick-interface-types" }
+wick-interface-types = { workspace = true }
 thiserror = { workspace = true }
-walrus = "0.19"
-data-encoding = "2.3.2"
-ring = "0.16"
-nuid = "0.4"
+wascap = { workspace = true }
+walrus = { workspace = true }
+data-encoding = { workspace = true }
+ring = { workspace = true }
+nuid = { workspace = true }
 tracing = { workspace = true }
 
 [dev-dependencies]
-anyhow = { version = "1.0" }
+anyhow = { workspace = true }
